Rename regression helper module
